requirements.txt
tensorflow
opencv-python
numpy
matplotlib
seaborn
scikit-learn
pillow
flask
#e-